Move WorkPool.js to services dir
Since it’s general purpose and not bolt specific
