+++ src/shared/services/WorkPool.js
