+++ src/shared/services/bolt/bolt.js
@@ -20,7 +20,7 @@
 
 import { v4 } from 'uuid'
 import { v1 as neo4j } from 'neo4j-driver-alias'
-import WorkPool from './WorkPool'
+import WorkPool from '../WorkPool'
 import * as mappings from './boltMappings'
 import * as boltConnection from './boltConnection'
 import {
